feat(core): Phase 4.0 Step 8 — audit fixes + ConversationStorePort wiring + REST API (#165)
Audit fixes (5 bugs):
- ConversationEvent token fields widened u32→u64 (silent truncation)
- create_run SQL uses separate created_at param (was duplicate ?5)
- delete_session returns false when session didn't exist
- PRAGMA foreign_keys=ON enforces ON DELETE CASCADE
- InboundEnvelope message IDs use UUID v4 (was collision-prone)

ConversationStorePort extended with 4 methods:
- update_label, update_goal, increment_message_count, add_token_usage
- ChatDbConversationStore adapter implements all 4

Gateway wiring:
- ConversationStorePort created from ChatDb in run_gateway()
- Added to AppState (Option<Arc<dyn ConversationStorePort>>)

REST API (3 new endpoints):
- GET /api/conversations — list sessions with optional prefix filter
- GET /api/conversations/:key — get session + recent events
- DELETE /api/conversations/:key — delete session

40 tests passing in fork_core/fork_adapters (was 36).

+++ src/fork_core/domain/conversation.rs
@@ -95,9 +95,9 @@ pub struct ConversationEvent {
     /// Run ID linking events to a single agent execution.
     pub run_id: Option<String>,
     /// Input tokens consumed (for LLM events).
-    pub input_tokens: Option<u32>,
+    pub input_tokens: Option<u64>,
     /// Output tokens produced (for LLM events).
-    pub output_tokens: Option<u32>,
+    pub output_tokens: Option<u64>,
     /// Event timestamp (unix seconds).
     pub timestamp: u64,
 }
